# Optimize RouteTableFactory.DetectAmbiguous routes