--- a/src/Components/Components/src/Routing/RouteTableFactory.cs
+++ b/src/Components/Components/src/Routing/RouteTableFactory.cs
@@ -6,14 +6,14 @@ using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using Microsoft.AspNetCore.Components.Routing;
 using Microsoft.AspNetCore.Routing;
-using Microsoft.AspNetCore.Routing.Template;
+using Microsoft.AspNetCore.Routing.Constraints;
 using Microsoft.AspNetCore.Routing.Patterns;
+using Microsoft.AspNetCore.Routing.Template;
 using Microsoft.AspNetCore.Routing.Tree;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using static Microsoft.AspNetCore.Internal.LinkerFlags;
-using Microsoft.AspNetCore.Routing.Constraints;
 
 namespace Microsoft.AspNetCore.Components;
 
@@ -200,63 +200,27 @@ internal class RouteTableFactory
 
     private static void DetectAmbiguousRoutes(TreeRouteBuilder builder)
     {
+        var seen = new Dictionary<InboundRouteEntry, InboundRouteEntry>(new InboundRouteEntryEqualityComparer());
+        seen.EnsureCapacity(builder.InboundEntries.Count);
+
         for (var i = 0; i < builder.InboundEntries.Count; i++)
         {
-            var left = builder.InboundEntries[i];
-            for (var j = i + 1; j < builder.InboundEntries.Count; j++)
-            {
-                var right = builder.InboundEntries[j];
-                var leftText = left.RoutePattern.RawText!.Trim('/');
-                var rightText = right.RoutePattern.RawText!.Trim('/');
-                if (left.Precedence != right.Precedence)
-                {
-                    continue;
-                }
+            var current = builder.InboundEntries[i];
 
-                var ambiguous = CompareSegments(left, right);
-                if (ambiguous)
-                {
-                    throw new InvalidOperationException($@"The following routes are ambiguous:
-'{leftText}' in '{left.Handler.FullName}'
-'{rightText}' in '{right.Handler.FullName}'
-");
-                }
+            if (seen.TryGetValue(current, out var existing))
+            {
+                var existingText = existing.RoutePattern.RawText!.Trim('/');
+                var currentText = current.RoutePattern.RawText!.Trim('/');
+                throw new InvalidOperationException($"""
+                    The following routes are ambiguous:
+                    '{existingText}' in '{existing.Handler.FullName}'
+                    '{currentText}' in '{current.Handler.FullName}'
+
+                    """);
             }
+
+            seen.Add(current, current);
         }
-    }
-
-    private static bool CompareSegments(InboundRouteEntry left, InboundRouteEntry right)
-    {
-        var ambiguous = true;
-        for (var k = 0; k < left.RoutePattern.PathSegments.Count; k++)
-        {
-            var leftSegment = left.RoutePattern.PathSegments[k];
-            var rightSegment = right.RoutePattern.PathSegments[k];
-            if (leftSegment.Parts.Count != rightSegment.Parts.Count)
-            {
-                ambiguous = false;
-                break;
-            }
-
-            for (var l = 0; l < leftSegment.Parts.Count; l++)
-            {
-                var leftPart = leftSegment.Parts[l];
-                var rightPart = rightSegment.Parts[l];
-                if (leftPart is RoutePatternLiteralPart leftLiteral &&
-                    rightPart is RoutePatternLiteralPart rightLiteral &&
-                    !string.Equals(leftLiteral.Content, rightLiteral.Content, StringComparison.OrdinalIgnoreCase))
-                {
-                    ambiguous = false;
-                    break;
-                }
-            }
-            if (!ambiguous)
-            {
-                break;
-            }
-        }
-
-        return ambiguous;
     }
 
     private static HashSet<string> GetParameterNames(RoutePattern routeTemplate)
@@ -333,5 +297,71 @@ internal class RouteTableFactory
             0 => 0,
             _ => throw new InvalidOperationException("Invalid comparison result."),
         };
+    }
+
+    private class InboundRouteEntryEqualityComparer : IEqualityComparer<InboundRouteEntry>
+    {
+        public bool Equals(InboundRouteEntry? x, InboundRouteEntry? y)
+        {
+            if (x is null)
+            {
+                return y is null;
+            }
+
+            if (y is null)
+            {
+                return false;
+            }
+
+            if (x.Precedence != y.Precedence)
+            {
+                return false;
+            }
+
+            for (var k = 0; k < x.RoutePattern.PathSegments.Count; k++)
+            {
+                var leftSegment = x.RoutePattern.PathSegments[k];
+                var rightSegment = y.RoutePattern.PathSegments[k];
+                if (leftSegment.Parts.Count != rightSegment.Parts.Count)
+                {
+                    return false;
+                }
+
+                for (var l = 0; l < leftSegment.Parts.Count; l++)
+                {
+                    var leftPart = leftSegment.Parts[l];
+                    var rightPart = rightSegment.Parts[l];
+                    if (leftPart is RoutePatternLiteralPart leftLiteral &&
+                        rightPart is RoutePatternLiteralPart rightLiteral &&
+                        !string.Equals(leftLiteral.Content, rightLiteral.Content, StringComparison.OrdinalIgnoreCase))
+                    {
+                        return false;
+                    }
+                }
+            }
+
+            return true;
+        }
+
+        public int GetHashCode(InboundRouteEntry obj)
+        {
+            var hashCode = new HashCode();
+            hashCode.Add(obj.Precedence);
+
+            for (var i = 0; i < obj.RoutePattern.PathSegments.Count; i++)
+            {
+                var segment = obj.RoutePattern.PathSegments[i];
+                for (var j = 0; j < segment.Parts.Count; j++)
+                {
+                    var part = segment.Parts[j];
+                    if (part is RoutePatternLiteralPart literal)
+                    {
+                        hashCode.Add(literal.Content, StringComparer.OrdinalIgnoreCase);
+                    }
+                }
+            }
+
+            return hashCode.ToHashCode();
+        }
     }
 }