Make RouteTableFactory.DetectAmbiguous faster using a hash table after it was reported as slow in #55114

## Description

I've added a performance benchmark to the code with the routes found in https://github.com/IgniteUI/igniteui-blazor-examples as reported in original ticket. I am unsure as to whether that's appropriate please advise if I must change them.

**Before**

|           Method |     Mean |   Error |  StdDev |  Op/s |     Gen 0 |    Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|--------:|--------:|------:|----------:|---------:|------:|----------:|
| CreateRouteTable | 223.1 ms | 4.32 ms | 6.05 ms | 4.482 | 1000.0000 | 333.3333 |     - |    749 MB |


**After**
|           Method |     Mean |     Error |    StdDev |  Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| CreateRouteTable | 3.889 ms | 0.0629 ms | 0.0588 ms | 257.1 | 7.8125 |     - |     - |      6 MB |


Fixes #55114