@andrewjsaid 2024-06-27
With new benchmark (dynamically generated routes) on latest commit I have these numbers.

**Before**
```
|           Method |     Mean |   Error |  StdDev |  Op/s |     Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|--------:|--------:|------:|----------:|------:|------:|----------:|
| CreateRouteTable | 208.6 ms | 3.92 ms | 3.66 ms | 4.794 | 1000.0000 |     - |     - |    619 MB |
```

**After**
```
|           Method |     Mean |     Error |    StdDev |  Op/s |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
| CreateRouteTable | 5.226 ms | 0.1043 ms | 0.1356 ms | 191.3 | 7.8125 |     - |     - |      7 MB |
```